+++ README.md
@@ -1,3 +1,6 @@
+## Лабораторная работа 5
+# Выполнила Терехина Е.В.
+
 # Объектно-ориентированное программирование
 
 ### Внешняя декомпозиция
